requirements.txt
exif==1.6.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
plum-py==0.8.7
